done factorial

+++ session-1/task1/factorial.c
@@ -9,11 +9,11 @@
     long int f[20];
     for( int k=0; k<1; ++k) {
     f[k]=1;
-    printf("%f\n",(double)f[k]);
+    printf("%ld\n",f[k]);
     }
     for( int k=1; k<20; ++k ) {
     f[k] = f[k-1]*k;
-    printf("%f\n",(double)f[k]);
+    printf("%ld\n",f[k]);
      }
 
     return 0;
